+++ apps/router-e2e/__e2e__/05-use-dom/components/02-actions.tsx
@@ -20,6 +20,9 @@ export default function Page({
 
   return (
     <div className="bg-slate-100 rounded-xl flex flex-1 flex-col gap-8">
+      <p className="text-lg text-blue-900 font-medium">
+        Parent platform: {process.env.EXPO_DOM_OS}
+      </p>
       <p
         className="text-lg text-blue-900 font-medium"
         onClick={() => {